+++ lib/gdal/raster_band.rb
@@ -1,13 +1,8 @@
 # frozen_string_literal: true
 
 require 'narray'
-require 'numo/narray'
 require_relative '../gdal'
-require_relative 'raster_band_mixins/algorithm_extensions'
 require_relative 'raster_band_mixins/algorithm_methods'
-require_relative 'raster_band_mixins/coloring_extensions'
-require_relative 'raster_band_mixins/extensions'
-require_relative 'raster_band_mixins/io_extensions'
 require_relative 'major_object'
 
 module GDAL
@@ -15,10 +10,6 @@ module GDAL
     include MajorObject
     include GDAL::Logger
     include RasterBandMixins::AlgorithmMethods
-    include RasterBandMixins::AlgorithmExtensions
-    include RasterBandMixins::ColoringExtensions
-    include RasterBandMixins::Extensions
-    include RasterBandMixins::IOExtensions
 
     ALL_VALID = 0x01
     PER_DATASET = 0x02