+++ lib/ogr/geometry_types/curve.rb
@@ -18,7 +18,7 @@ module OGR
         FFI::OGR::API.OGR_G_GetZ(@c_pointer, point_number)
       end
 
-      # @param [Integer] Index of the point to get.
+      # @param number [Integer] Index of the point to get.
       # @return [Array<Float, Float, Float>] [x, y] if 2d or [x, y, z] if 3d.
       def point(number)
         x_ptr = FFI::MemoryPointer.new(:double)
@@ -34,21 +34,6 @@ module OGR
         end
       end
       alias get_point point
-
-      # It seems as if {{#point}} should return an OGR::Point, but since OGR's
-      # OGR_G_GetPoint only returns coordinates, this allows getting the point
-      # as an OGR::Point.
-      #
-      # @param [Integer] Index of the point to get.
-      # @return [OGR::Point]
-      # TODO: Move to an extension.
-      def point_geometry(number)
-        coords = point(number)
-        point = OGR::Point.new
-        point.set_point(0, *coords)
-
-        point
-      end
 
       # Adds a point to a LineString or Point geometry.
       #
@@ -73,24 +58,6 @@ module OGR
         else
           FFI::OGR::API.OGR_G_SetPoint_2D(@c_pointer, index, x, y)
         end
-      end
-
-      # @return [Enumerator]
-      # @yieldparam [OGR::Point]
-      # TODO: Move to an extension.
-      def each_point_geometry
-        return enum_for(:each_point_as_geometry) unless block_given?
-
-        point_count.times do |point_num|
-          yield point_as_geometry(point_num)
-        end
-      end
-
-      # @return [Array<OGR::Point>]
-      # @see {{#each_point_geometry}}, {{#point_geometry}}
-      # TODO: Move to an extension.
-      def point_geometries
-        each_point_geometry.to_a
       end
 
       # @return [Array<Array<Float>>] An array of (x, y) or (x, y, z) points.
@@ -124,17 +91,6 @@ module OGR
       alias get_points points
       alias point_values points
 
-      # @param geo_transform [GDAL::GeoTransform]
-      # @return [Array<Array>]
-      def pixels(geo_transform)
-        log "points count: #{point_count}"
-        points.map do |x_and_y|
-          result = geo_transform.world_to_pixel(*x_and_y)
-
-          [result[:pixel].to_i.abs, result[:line].to_i.abs]
-        end
-      end
-
       # @param new_count [Integer]
       def point_count=(new_count)
         FFI::OGR::API.OGR_G_SetPointCount(@c_pointer, new_count)
@@ -146,18 +102,6 @@ module OGR
       # @return [Float] 0.0 for unsupported geometry types.
       def length
         FFI::OGR::API.OGR_G_Length(@c_pointer)
-      end
-
-      def start_point
-        point(0)
-      end
-
-      def end_point
-        point(point_count - 1)
-      end
-
-      def closed?
-        start_point == end_point
       end
     end
   end