+++ spec/unit/gdal/dataset_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
 require 'gdal/dataset'
 
 RSpec.describe GDAL::Dataset do
@@ -102,7 +101,7 @@ RSpec.describe GDAL::Dataset do
   describe '#projection' do
     let(:expected_wkt) do
       'GEOGCS["unknown",DATUM["unknown",SPHEROID["Bessel 1841",' \
-        '6377397.155,299.1528128000008,AUTHORITY["EPSG","7004"]],' \
+        '6377397.155,299.1528128000033,AUTHORITY["EPSG","7004"]],' \
         'TOWGS84[598.1,73.7,418.2,0.202,0.045,-2.455,6.7]],' \
         'PRIMEM["Greenwich",0],UNIT["degree",0.0174532925199433]]'
     end