+++ spec/unit/gdal/raster_attribute_table_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
 require 'gdal/raster_attribute_table'
 
 RSpec.describe GDAL::RasterAttributeTable do