+++ spec/unit/gdal/rpc_info_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
 require 'gdal/rpc_info'
 
 RSpec.describe GDAL::RPCInfo do