+++ spec/unit/ogr/geometries/point_spec.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
 require 'ogr/geometry'
 
 RSpec.describe OGR::Point do
@@ -237,7 +236,7 @@ RSpec.describe OGR::Point do
     let(:wkt) { 'POINT EMPTY' }
 
     describe '#coordinate_dimension' do
-      specify { expect(subject.coordinate_dimension).to eq 0 }
+      specify { expect(subject.coordinate_dimension).to eq 2 }
     end
 
     describe '#geometry_count' do
@@ -299,8 +298,8 @@ RSpec.describe OGR::Point do
     end
 
     describe '#envelope' do
-      it 'returns nil' do
-        expect(subject.envelope).to be_nil
+      it 'returns an OGR::Envelope' do
+        expect(subject.envelope).to be_a OGR::Envelope
       end
     end
   end